refactor(embedding): remove the rune-based embedding plan path
Embedding inputs now derive from normalized evidence through the
tokenizer-specific generator in the document package. The older plan,
recipe, and distillation flow chunked by rune count over the legacy
normalized document and had no consumer outside its own package. Keeping
both would leave every later change to explain which chunker is
authoritative.

The egress and vector-space identities and the shared retrieval policy
stay, since deployment config and search still use them. Normalized
document chunks remain only for non-embedding uses.
