+++ document/embedding/doc.go
@@ -1,5 +1,7 @@
-// Package embedding prepares deterministic, provider-neutral document inputs
-// for text embedding and optional pre-embedding distillation.
+// Package embedding holds provider-neutral egress identities and shared
+// retrieval policy for document embeddings. Tokenizer-specific document
+// inputs are generated by document.BuildEmbeddingInputs from normalized
+// evidence.
 //
 // The package performs no network, storage, queue, database, or consent work.
 // Applications own those effects and use the fingerprints returned here as
